Move JSON format tests to their own file.
